+++ src/components/timeline/timeline.vue
@@ -16,6 +16,7 @@
       >
         {{ $t('timeline.up_to_date') }}
       </div>
+      <TimelineQuickSettings v-if="!embedded" />
     </div>
     <div :class="classes.body">
       <div
@@ -103,9 +104,12 @@
   max-width: 100%;
   flex-wrap: nowrap;
   align-items: center;
+  position: relative;
+
   .loadmore-button {
     flex-shrink: 0;
   }
+
   .loadmore-text {
     flex-shrink: 0;
     line-height: 1em;