+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Added reason field for registration when approval is required
+- Added a quick settings to timeline header for easier access
 - Added option to mark posts as sensitive by default
 
 ## [2.2.3] - 2021-01-18